move rabbit mq configuration to seperate class
